Support discriminated unions correctly
This makes use of the discriminated union support in the xdr2 package,
and uses that to correct several protocol errors in the previous code.

This client is still woefully incomplete and likely does not handle
various corner cases correctly.

Basic testing performed against Linux kernel NFS server and NetApp
simulator.